(PC-12976)[API] refactor: move create_orphan_dms_application to subscription repository
